Get shit done (#23)
* Finished Abhay's Work

Implemented Time skipping of 3 hours
Added trackers for time and date ranges
Prevented time looping
Configured date select on bottom of itinerary
Displayed week range on top of week itinerary

* Itinerary Display (no day overlap)

Itinerary Display now works off of bufferItinerary, Next must implement the left itinerary bar.

Works for both the single and week itinerary

* Itinerary Left Panel now works, Figured out conflicts

The itinerary implementation is now close to being done, with the exception of filters.

Conflict only appears when two events are trying to be displayed at once, it does not detect overlapping events due to computational restrictions. Can be upgraded at a later time, but good for now.

* Added Custom Events, Fixed Some Display Issues

Custom Events now implemented.

Added border around each event (top/bottom)

Fixed logic issues with conflicts and displays where an extra span was taken.

Added reduced display for quick events

* Implemented Itinerary Loading

Itinerary Loading now works as expected. Current error with loading message displaying, but that's minor and not necessarily needed.

Also fixed custom event time calculation, now midnight is 0:00 as required.

* Delete Itinerary Implemented

Allows user to delete from templates saved on their account.

* Configured User Logged In Banner

Banner now accurately reflects whether the user is logged in.

Also hides the delete button when no user is logged in.

* Create New Itinerary Implemented

Now button wipes current buffer

Renamed ThreeFrame buttons to better reflect what they do.

* Print Window Implemented

This window doesn't actually do anything. It has error checking within but ultimately ends with a made up system printing dialog being called.

* Commented Testing Populators, Fixed issue with Itinerary Left

Commented out events hardcoded for testing .

Fixed 12:00am issue with itinerary Left

+++ FlamePlanner/Itinerary_leftpannel.xaml.cs
@@ -21,21 +21,75 @@ namespace FlamePlanner
     public partial class Itinerary_leftpannel : Page
     {
         private MainWindow mw;
-        public Itinerary_leftpannel(MainWindow mw)
+        private Itinerarypage ItinPage;
+        public Itinerary_leftpannel(MainWindow mw, Itinerarypage ip)
         {
-            InitializeComponent();
             this.mw = mw;
+            this.ItinPage = ip;
+            InitializeComponent();
+            populateEventPanel();
+
+            if (mw.loggedIn)
+            {
+                accountBanner.Text = "Username: " + mw.currentAcount;
+                accountBanner.Background = Brushes.LightBlue;
+            }
+            else
+            {
+                accountBanner.Text = "Not Logged In";
+                accountBanner.Background = Brushes.LightYellow;
+                deleteButton.Visibility = Visibility.Hidden;
+            }
         }
 
-        private void CheckBox_Checked(object sender, RoutedEventArgs e)
+        public void populateEventPanel()
         {
+            eventPanel.Children.Clear();
+            foreach (EventObject e in mw.bufferItinerary.eventList)
+            {
+                EventItineraryBanner eib = new EventItineraryBanner(mw, ItinPage, this, e);
+                eventPanel.Children.Add(eib);
+            }
+        }
 
+        /// <summary>
+        /// This function ensures checks match the visibility of each event, Could occur when a conflict is detected
+        /// </summary>
+        public void redoCheckboxSelections()
+        {
+            foreach (object o in eventPanel.Children)
+            {
+                if (o != null && o.GetType() == typeof(EventItineraryBanner))
+                {
+                    EventItineraryBanner e = o as EventItineraryBanner;
+                    e.setCheck(e.eventO.isVisible); //sets check to reflect whether it should be visible or not
+                }
+            }
         }
 
         private void deleteButton_Click(object sender, RoutedEventArgs e)
         {
-            deleteConfirmation dc = new deleteConfirmation(mw);
-            dc.ShowDialog();
+            //Load another itinerary (Have they logged in?) Prompt to save?
+            if (!mw.loggedIn)
+            {
+                itinerarySaveOptions isow = new itinerarySaveOptions(mw);
+                isow.ShowDialog();
+            }
+            else //logged in
+            {
+                Account acc = mw.AccountDatabase[mw.currentAcount];
+                if (acc.itineraryDict.Count == 0)
+                {
+                    MessageBox.Show("There are No Itinearies Saved on this Account to Delete");
+                }
+                else
+                {
+                    deleteConfirmation dc = new deleteConfirmation(mw);
+                    dc.ShowDialog();
+                }
+            }
+
+            
         }
     }
 }
